fix: change calories per reps to double

+++ src/models/workoutDataModel.js
@@ -15,7 +15,7 @@ WorkoutData.init(
             allowNull: false,
         },
         calories_per_reps: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DOUBLE(5, 2),
             allowNull: false,
         },
         bicep: {
